requirements.txt
# Python 3.11+ recommended for optimal performance with AI/ML workloads
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
python-dotenv==1.0.0
httpx==0.27.0

# LangChain and LangGraph for ReAct agent
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-anthropic==0.1.0
langchain-community==0.0.10
tavily-python==0.3.1

# Additional dependencies to resolve conflicts
pillow>=8.0.0,<12.0.0
pyarrow>=7.0.0
grpcio>=1.48.2,<2.0
markdown>=2.6.8
